pyproject.toml
[project]
name = "ai-data-insights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.4",
    "openai>=2.21.0",
    "pandas>=2.3.3",
    "playwright>=1.58.0",
    "streamlit>=1.54.0",
    "google-generativeai",
    "yfinance",
    "fastapi",
    "uvicorn",
]
